LayoutTests/picture/expectedfails/picture-form-image-access-by-name.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>

<form>
<span id="testSpan">
<picture id="1" name="imageElement"></picture>
</span>
</form>

<div id="console"></div>
<script>
function debugLog(s) {
    var console = document.getElementById("console");
    console.appendChild(document.createTextNode(s));
    console.appendChild(document.createElement("br"));
}

var form = document.forms[0];
var span = document.getElementsByTagName('span')[0];
shouldBe("form.imageElement.id", "'1'");

span.innerHTML = "<picture id='2' name='imageElement'></picture>";
shouldBe("form.imageElement.id", "'2'");

span.innerHTML = "<picture id='2' name='newImage'></picture>";

// imageElement has been removed from the DOM, but,
// IE 6 has a quirk where once accessed, form images are always accessible by name
shouldBeTrue("!!form.imageElement");

// This quirk has no bearing on document. access
shouldBeUndefined("document.imageElement");

// Of course, the new image should be accessible both ways.
shouldBeTrue("!!document.newImage");
shouldBeTrue("!!form.newImage");

// However, if you never access an image element by name it does not stick around after being removed:
span.innerHTML = "<picture id='3' name='thirdImage'></picture>";
span.innerHTML = "<picture id='4' name='fourthImage'></picture>";
shouldBeUndefined("form.thirdImage");
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
